+++ lib/shared/models/cripto_model.dart
