+++ lib/wallet/widgets/hide_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../shared/template/wallet_providers.dart';
+import '../../shared/providers/wallet_providers.dart';
 
 class HideButton extends HookConsumerWidget {
   const HideButton({
