+++ lib/wallet/widgets/total_value.dart
@@ -1,9 +1,9 @@
-import 'package:crypto_listing/shared/template/wallet_providers.dart';
+import '../../shared/providers/wallet_providers.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../shared/template/number_formatter.dart';
+import '../../shared/widgets/number_formatter.dart';
 
 class TotalValue extends StatefulHookConsumerWidget {
   final double totalReais;
